Add more integ tests for query insights (#71)
(cherry picked from commit 2c02a2d22a7a28e8b493711f2d7ff9b571274298)
