+++ build.gradle
@@ -263,6 +263,7 @@ integTest {
   if (System.getProperty("security.enabled") == "true" || System.getProperty("https") == "true") {
     // Exclude this IT, because they executed in another task (:integTestWithSecurity)
     exclude 'org/opensearch/plugin/insights/rules/resthandler/top_queries/TopQueriesRestIT.class'
+    exclude 'org/opensearch/plugin/insights/core/exporter/QueryInsightsExporterIT.class'
   }
 }
 
@@ -333,6 +334,8 @@ task integTestWithSecurity(type: RestIntegTestTask) {
     jvmArgs '-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=*:5005'
   }
 
+  // NOTE: this IT config discovers only junit5 (jupiter) tests.
+  // https://github.com/opensearch-project/sql/issues/1974
   filter {
     includeTestsMatching 'org.opensearch.plugin.insights.rules.resthandler.top_queries.TopQueriesRestIT'
   }
